theme: Fit page to viewport
Fit page to the browser's viewport. Content will be made to expand /
shrink as necessary. Also, include an element's border in its box model
sizing, as I feel it's more sensible when I want to maintain a specific
size for an element while using a border.

Each page will fit within a `<section>` element, and the `slide` class
can now be placed on this element directly.

Refs https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing