feat(nav): Add styling for left nav
When the user navigates to a new page, the transition should reflect the
direction to which the navigation has occurred. I'm still considering a
few methods as to how to retrieve the previous and next routes to
implement this into code, however. A watcher on the $route property will
bw required, so I've added the initial code for that in the meantime.

A nicesty with the Vue.js router is that it automatically applies the
`router-link-active` class on the currently active `<router-link>`, so
we can set the styling for that essentially for free.

Refs https://router.vuejs.org/en/api/router-link.html#props