feat(nav): Make slide transition dynamic
Bind the transition to a data property that is dynamically set to either
`slide-right` or `slide-left` based on which direction to user is
navigating with respect to "page" number (i.e. order of the links on the
navigation bar).

I've also refactored the navigation menu to generate the list of links
from the supplied `routes`. Each route will now also be named, shown as
its respective link text, and routes without names will be ignored (in
order to avoid adding a list item for the '*' path.