refactor(nav): Don't generate navigation links
Automatically generating the navigation links from the routes array has
some limitations, such as hardcoding wildcard routes (in some way) to be
an exception. This also means that child routes and other complex
navigation will not be possible.

The existing solution also required a route name to use as the link
title, but that was likely a misuse of the available feature.