style: Style codebase against recommended configs
Use ESLint and the Vue.js plugin for ESLint to catch potential issues
and to style the codebase in a consistent manner. I'm choosing to simply
use the recommended configuration of both ESLint itself as well as the
Vue.js plugin.

The Vue.js plugin provides better config options than the defaults, such
as the `module` source type and an updated ECMAScript version. This
means I can get away with a bare ESLint config file.

One thing of note is that, while the version of the Vue.js is currently
beta, this will not be a habit as I will in the future prefer "final"
versions. However, this case is acceptable as it allows everything to
work nicely over the current "final" version available.