refactor(nav): Use `meta` prop to order routes
An `index` meta property is a much more elegant solution to the problem
of assigning some sort of order to a route, as it's the official way to
assign extra information anyway, and there's no need to use any
array-based functions to calculate the index. This is more explicit,
which is likely a good thing, though if the order of the routes changes
then two places now need to be updated.