chore(package): Remove superfluous `os` config
Travis CI automatically defaults to Linux containers for testing, so
there's no need to explicitely include the config for Linux as a test
platform. This package will only ever be built on my own machine, which
will likely never be an Apple product, and will likely only ever be
deployed on a Linux system.