refactor(nav): Simplify transition styling
If using the 'out-in' transition mode, there's no need to specify the
position as absolute for each page / slide, and thus also no need to
specify the parent position as relative.

While we're tweaking the code, we can fine-tune it a bit more so that
the pag / slide entering will ease out and the page / slide leaving will
ease in.